Renombre de formato de java script a jsx para mostrar que es REACT
